fix: incorrect timezone handling in observations pane
I noticed that the temperature observations pane had a timezone handling error resulting in data for UTC being displayed as local time. After investigating the return values from get_obs(), I decided the best option would be to get all observations in UTC, then convert to the local time of the station for display. This required adding a timezone attribute to the Location class and using tzfpy to look up the correct zone based on lat/long. Weather observations are displayed in the correct local time of the station as a result of these changes.